MSH work on home page
I'm not entirely happy with how the swap diagrams are looking.  I am using excel to create the picture, saving it to word as a picture, then saving it as a png file.  The lettters in the png file are darker and fuzzier than on the original excel sheet.  Do you know any better way to do this?